full_requirements.txt
absl-py==2.1.0
actionlib==1.13.2
angles==1.9.13
beautifulsoup4==4.12.3
bondpy==1.8.6
box2d-py==2.3.5
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.0
cycler==0.12.1
decorator==4.4.2
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
docker-pycreds==0.4.0
dynamic-reconfigure==1.7.1
etils==1.7.0
exceptiongroup==1.2.1
Farama-Notifications==0.0.4
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.5.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gdown==5.2.0
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
grpcio==1.64.0
gymnasium==0.28.1
gymnasium-robotics==1.2.2
idna==3.7
image-geometry==1.16.0
imageio==2.34.1
imageio-ffmpeg==0.5.0
importlib_resources==6.4.0
iniconfig==2.0.0
interactive-markers==1.12.0
jax-jumpy==1.0.0
Jinja2==3.1.4
joint_state_publisher==1.15.0
joint_state_publisher_gui==1.15.0
kiwisolver==1.4.5
laser_geometry==1.6.7
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
message-filters==1.15.13
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.3
networkx==3.3
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
# Editable install with no version control (omnirobust==0.1.0)

packaging==24.0
pandas==2.0.3
pettingzoo==1.24.3
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
proglog==0.1.10
protobuf==4.25.3
psutil==5.9.8
pygame==2.1.3
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.1
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2024.1
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
resource_retriever==1.12.6
rich==13.7.1

rosbag==1.15.13
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.13
roslaunch==1.15.13
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.13
rosmake==1.15.8
rosmaster==1.15.13
rosmsg==1.15.13
rosnode==1.15.13
rosparam==1.15.13
rospy==1.15.13
rosservice==1.15.13
rostest==1.15.13
rostopic==1.15.13
rosunit==1.15.8
roswtf==1.15.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.2
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.2
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.11
seaborn==0.13.2
sensor-msgs==1.13.1
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
smmap==5.0.1
soupsieve==2.5
swig==4.2.1
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
tomli==2.0.1
topic-tools==1.15.13
torch==2.3.0
tqdm==4.66.4
triton==2.3.0
typer==0.12.3
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
Werkzeug==3.0.3
xacro==1.14.10
xmltodict==0.13.0
zipp==3.19.1
